+++ packages/datasource-sql/test/datasource-factory.integration.test.ts
@@ -199,107 +199,7 @@ const RELATION_MAPPING = {
   },
 };
 
-class MyBuilder implements Builder {
-  models: { [p: string]: Model };
-  logger?: Logger;
-
-  buildDataSource(): DataSource {
-    return undefined;
-  }
-
-  defineModel(): Promise<void> {
-    throw new Error();
-  }
-
-  defineRelation(): Promise<void> {
-    return Promise.resolve();
-  }
-
-  getRelatedTables(): Promise<string[]> {
-    return Promise.resolve([]);
-  }
-
-  getTableNames(): Promise<string[]> {
-    return Promise.resolve([]);
-  }
-}
-
-describe('SqlDataSourceFactory', () => {
-  describe('when defining a model throw an error', () => {
-    it('should display a log message and does not create its relation', async () => {
-      const myOrm = new MyBuilder();
-      myOrm.defineModel = jest.fn().mockImplementation(() => {
-        throw new Error('Error Message');
-      });
-      myOrm.logger = jest.fn();
-      myOrm.getTableNames = jest.fn().mockReturnValue(['MyTable']);
-      myOrm.defineRelation = jest.fn();
-      myOrm.getRelatedTables = jest.fn().mockReturnValue(['MyTable']);
-
-      await SqlDataSourceFactory.build(myOrm);
-
-      expect(myOrm.logger).toHaveBeenCalledWith(
-        'Warn',
-        'Skipping table "MyTable" and its relations because of error: Error Message',
-      );
-      expect(myOrm.defineRelation).not.toHaveBeenCalled();
-    });
-
-    describe('when there are many tables and a model with an error', () => {
-      it('creates all the models and the relations when there is no error', async () => {
-        const myOrm = new MyBuilder();
-        myOrm.logger = jest.fn();
-        myOrm.defineRelation = jest.fn();
-        myOrm.getTableNames = jest.fn().mockReturnValue(['TableWithError', 'TableWithNoError']);
-        myOrm.defineModel = jest.fn().mockImplementation((tableName: string) => {
-          if (tableName === 'TableWithError') {
-            throw new Error('Error Message');
-          }
-        });
-        myOrm.getRelatedTables = jest.fn().mockImplementation((tableName: string) => {
-          if (tableName === 'TableWithError') {
-            return ['TableWithError'];
-          }
-
-          return ['TableWithNoError', 'AndOtherTable'];
-        });
-
-        await SqlDataSourceFactory.build(myOrm);
-
-        expect(myOrm.logger).toHaveBeenCalledTimes(1);
-
-        expect(myOrm.defineRelation).toHaveBeenCalledWith('TableWithNoError');
-        expect(myOrm.defineRelation).toHaveBeenCalledTimes(1);
-      });
-
-      describe('when a related table has an error', () => {
-        it('should does not create the relation', async () => {
-          const myOrm = new MyBuilder();
-          myOrm.logger = jest.fn();
-          myOrm.defineRelation = jest.fn();
-          myOrm.getTableNames = jest.fn().mockReturnValue(['TableWithError', 'TableWithNoError']);
-          myOrm.defineModel = jest.fn().mockImplementation((tableName: string) => {
-            if (tableName === 'TableWithError') {
-              throw new Error('Error Message');
-            }
-          });
-          myOrm.getRelatedTables = jest.fn().mockImplementation((tableName: string) => {
-            if (tableName === 'TableWithError') {
-              return ['TableWithError'];
-            }
-
-            return ['TableWithNoError', 'TableWithError'];
-          });
-
-          await SqlDataSourceFactory.build(myOrm);
-
-          expect(myOrm.logger).toHaveBeenCalledTimes(1);
-          expect(myOrm.defineRelation).toHaveBeenCalledTimes(0);
-        });
-      });
-    });
-  });
-
+describe('SqlDataSourceFactory > Integration', () => {
   describe('when the the table has an "id" without primary key constraint', () => {
     const setupDatabase = async (database: string): Promise<Sequelize> => {
       let sequelize: Sequelize;
